# openapi.yaml
openapi: 3.0.0
info:
  title: Student Management System API
  version: 1.0.0
  description: |
    Comprehensive API documentation for managing students, institutes, courses, and academic results
servers:
  - url: http://localhost:3000/api
    description: Development server
components:
  parameters:
...
    courseIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        lastLogin:
          type: string
          format: date-time
...
    Course:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
        description:
          type: string
        credits:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        status_code:
          type: integer
        status:
          type: string
        message:
          type: string
        error:
          type: string
paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  example: password123
      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      user:
                        $ref: "#/components/schemas/User"
                      token:
                        type: string
        400:
          description: Validation error or user exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  example: password123
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      user:
                        $ref: "#/components/schemas/User"
                      token:
                        type: string
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        200:
          description: Current user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      user:
                        $ref: "#/components/schemas/User"
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /courses:
    get:
      tags: [Courses]
      summary: List all courses
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        200:
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      courses:
                        type: array
                        items:
                          $ref: "#/components/schemas/Course"
                      pagination:
                        type: object
                        properties:
                          total:
                            type: integer
                          page:
                            type: integer
                          pages:
                            type: integer
    post:
      tags: [Courses]
      summary: Create new course
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Course"
      responses:
        201:
          description: Course created
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      course:
                        $ref: "#/components/schemas/Course"
  /courses/{id}:
    get:
      tags: [Courses]
      summary: Get course by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/courseIdParam"
      responses:
        200:
          description: Course details
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      course:
                        $ref: "#/components/schemas/Course"
        404:
          description: Course not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      tags: [Courses]
      summary: Update course
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/courseIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Course"
      responses:
        200:
          description: Course updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  data:
                    type: object
                    properties:
                      course:
                        $ref: "#/components/schemas/Course"
        404:
          description: Course not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      tags: [Courses]
      summary: Delete course
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/courseIdParam"
      responses:
        200:
          description: Course deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status_code:
                    type: integer
                  status:
                    type: string
                  message:
                    type: string
        404:
          description: Course not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...